refactor: shared가 pages를 참조하던 모듈을 제 위치로 이동
