+++ src/pages/home/analytics/resolveShopTriggerContext.ts
