+++ src/shared/constants/funnel.ts
