+++ src/shared/detection/hooks/useFurnitureHotspots.ts
@@ -9,8 +9,6 @@ import { useCallback, useEffect, useMemo, useReducer, useRef } from 'react';
 //   - 기준: cabinet은 refine confidence, 그 외는 모델 score 사용(단일 랭크 스코어로 비교)
 //   - 방식: 신뢰도/점수 상위 K개만 노출, K는 FALLBACK_MAX_CANDIDATES
 
-import { runSerializedInferenceTask } from '@pages/generate/utils/inferenceTaskScheduler';
-
 import { OBJ365_MODEL_PATH } from '@shared/detection/constants';
 import { useONNXModel } from '@shared/detection/hooks/useOnnxModel';
 import type { ProcessedDetections } from '@shared/detection/types';
@@ -18,6 +16,7 @@ import {
   logFurniturePipelineEvent,
   reportFurniturePipelineWarning,
 } from '@shared/detection/utils/furniturePipelineMonitor';
+import { runSerializedInferenceTask } from '@shared/detection/utils/inferenceTaskScheduler';
 
 import {
   buildHotspotsPipeline,
