+++ src/shared/detection/utils/inferenceTaskScheduler.ts
